Add 2nd commit calculation in LDAA Sign Commit
This code will cause a seg fault due to the size of the B NTT polynomial
matrix. The commit sign command must receive the B polynomial matrix to
calculate the commit, however, this matrix requires 2.5GB(!). Since the
TPM simulator operates solely on stack based memory, and, generally,
a stack size in a standard machine is aroung 8192kB, an allocation of this
magnitude will inevitably overflow the stack and cause a segmentation
fault. This issue will be fixed in a future commit.