Addresses half of the memory issues discussed
The memory required by the Sign Commit command was reduced in two ways:
1 - The B NTT matrix is only composed of the 0th order coefficient of
    each polynomial. This is because the commit operation would multiply
    all of the coefficients [1-255] in the matrix by zero.

    Therefore, the size of the B NTT matrix that must be passed to the
    TPM reduces from 2.5GB to 9.9MB, in the worst case.

2 - The secret randomness matrix was also downsized from a matrix of
  polynomials to a matrix of scalars. All of the functions which operate
  over these types have been updated to reflect the changes described.

The sign state of the commit now resides inside the TPM NV memory,
however, this solution is not perfect. The sign state requires 16MB of
memory which may be too large for the NV. I'll probably revisit this
issue in a later commit, currently it's not a blocker.