Update TPM to support LDAA
This commit is very big and I apologize for it.

Here are the key points of this commit:
- All structs from the TPM2B family were update to
support sizes of 32-bit. Relevant files are:
-- CommandDispatchData.h and CommandDispatcher.c: updated command descriptor
sizes from 16-bit to 32-bit
-- Marshal.c and Marshal_fp.h: updated marshaling functions
-- Unmarshal.c and Unmarshal_fp.h: updated marshaling functions
-- TPMB.h and TpmTypes.h: for the updated TPM2B structs

- TcpServer.c and TcpServerPosix.c: increase input and output buffers
from 1MiB to 64MiB.

- AsymmetricCommands.c: Fix bug in Kyber 2Phase key exchange. Alice's
public static key wasn't being passed to the TPM (also relevant
Kyber_2Phase_KEX_fp.h). (I know this should be in a separate commit. I'm
sorry). Update LDAA Sign Commit command user API: the user now has to call the
TPM for each commit it must process due to memory limitations and to keep the
TPMs memory footprint low.

- CryptLDaa.c, CryptLDaa_fp.h and LDaa_SignCommit_fp.h: Update SignCommit
function to an iterative interface instead of a bulk one. Move token link
processing to match the iterative interface. We expect the user to start the
commit processing by genrating a token link and then proceed to calculate the
commit in whatever order they see fit.

- Global.c/.h: Increase actionIoBuffer from 8kiB to 16kiB.

- Implementation.h:
-- Increase number of object handles in the TPM from 3 to 5;
-- Increase context size from 8kiB to 64kiB;
-- Increase command size from 16kiB to 64kiB;
-- Increase response size from 16kiB to 64MiB;
-- Increase NV memory size from 65kiB to 20MiB