Add support for SHA3
Move all hashing functions of Kyber and Dilithium to the internal
CryptHash functions. There is also limited support for the XOF functions
in SHA3: SHAKE128 and SHAKE256. These are not documented,
thus they are not expected to be used even though it is possible to send
certain magic bytes to obtain a hash. The maximum digest supported for
the XOFs is 1024 bytes.